Change double to float in ANNA program

+++ HideWordSolver/ANNA/Poc/backward.c
@@ -1,17 +1,17 @@
 #include "poc.h"
 
 void backward(int nb_letter,
-		double **w_output, // output_neuron x hidden_neuron
-		double **input, // input_neuron x nb_letter
-		double **hidden, // hidden_neuron x nb_letter
-		double **output, // output_neuron x nb_letter
-		double **expected_output, // output_neuron x nb_letter
-		double **output_error, // output_neuron x nb_letter
-		double **dw_output, // output_neuron x hidden_neuron
-		double *db_output, // output_neuron
-		double **hidden_error, // hidden_neuron x nb_letter
-		double **dw_input, // hidden_neuron x input_neuron
-		double *db_input) // hidden_neuron
+		float **w_output, // output_neuron x hidden_neuron
+		float **input, // input_neuron x nb_letter
+		float **hidden, // hidden_neuron x nb_letter
+		float **output, // output_neuron x nb_letter
+		float **expected_output, // output_neuron x nb_letter
+		float **output_error, // output_neuron x nb_letter
+		float **dw_output, // output_neuron x hidden_neuron
+		float *db_output, // output_neuron
+		float **hidden_error, // hidden_neuron x nb_letter
+		float **dw_input, // hidden_neuron x input_neuron
+		float *db_input) // hidden_neuron
 {
     // Apply back propagation on weights and biases between output and hidden layer
 	for (size_t i = 0; i < output_neuron; i++)
@@ -29,19 +29,19 @@ void backward(int nb_letter,
 	{
 		for (size_t j = 0; j < hidden_neuron; j++)
 		{
-			dw_output[i][j] = ((double)1 / nb_letter) * dw_output[i][j];
+			dw_output[i][j] = ((float)1 / nb_letter) * dw_output[i][j];
 		}
 	}
 
 	for (size_t i = 0; i < output_neuron; i++)
 	{
-		double sum = 0;
+		float sum = 0;
 		for (int j = 0; j < nb_letter; j++)
 		{
 			sum += output_error[i][j];
 		}
 		
-		db_output[i] = ((double)1 / nb_letter) * sum;
+		db_output[i] = ((float)1 / nb_letter) * sum;
 	}
 
 	poc_matrix_product_t(output_neuron, hidden_neuron, w_output, output_neuron,
@@ -64,18 +64,18 @@ void backward(int nb_letter,
 	{
 		for (size_t j = 0; j < input_neuron; j++)
 		{
-			dw_input[i][j] = ((double)1 / nb_letter) * dw_input[i][j];
+			dw_input[i][j] = ((float)1 / nb_letter) * dw_input[i][j];
 		}
 	}
 
 	for (size_t i = 0; i < hidden_neuron; i++)
 	{
-		double sum = 0;
+		float sum = 0;
 		for (int j = 0; j < nb_letter; j++)
 		{
 			sum += hidden_error[i][j];
 		}
 		
-		db_input[i] = ((double)1 / nb_letter) * sum;
+		db_input[i] = ((float)1 / nb_letter) * sum;
 	}
 }
