+++ HideWordSolver/ANNA/relu.c
@@ -1,4 +1,4 @@
-double relu(double d)
+float relu(float d)
 {
     return d > 0 ? d : 0.0;
 }
